requirements.txt
Cartopy==0.24.1
certifi==2025.7.14
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.0
geodatasets==2024.8.0
geopandas==1.1.1
idna==3.10
iniconfig==2.1.0
kiwisolver==1.4.8
matplotlib==3.10.3
numpy==2.2.6
numpy-quaternion==2024.0.10
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.1
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
Pygments==2.19.2
pyogrio==0.11.0
pyparsing==3.2.3
pyproj==3.7.1
pyshp==2.3.1
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
scipy==1.16.0
shapely==2.1.1
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
